move registerTableWithWatermark from StreamTableEnvironment to StreamTableTestUtil
